Just too many changes to list, check description.
I'll just be as brief as possible since there are way too many changes to recall.

* Finally added some sort of machine learning driven AI. It's still stupid as all hell and doesn't work properly whatsoever, but this is very, veeeery early stages.

* Added new particle effects.

* Added menus.

* Added new music.

* Added a new stage inspired by Medieval times.

* Added a new actor component for debugging attacks.

* Fixed a ridiculous amount of bugs... So. Many. Bugs. Finally over...

* Made adjustments to gameplay: You can no longer move away from center indefinetly. You'll eventually be stopped.

+++ Source/Valr/ANN/ArtificialNN.h
@@ -24,15 +24,15 @@ public:
 	ArtificialNN(int nIn, int nOut, int nHidden, int nPH, double a);
 	~ArtificialNN();
 
-	TArray<double> Train(TArray<double> inputValues, TArray<double> desiredOutput);
+	TArray<int> Train(TArray<int> inputValues, TArray<int> desiredOutput);
 
-	TArray<double> CalcOutput(TArray<double> inputValues);
+	TArray<int> CalcOutput(TArray<int> inputValues);
 
-	TArray<double> CalcOutput(TArray<double> inputValues, TArray<double> desiredOutput);
+	TArray<int> CalcOutput(TArray<int> inputValues, TArray<int> desiredOutput);
 
 	FString PrintWeights();
 
-	void UpdateWeights(TArray<double> outputs, TArray<double> desiredOutput);
+	void UpdateWeights(TArray<int> outputs, TArray<int> desiredOutput);
 
 	void LoadWeights( FString weightStr);
